+++ IcVibracoes.Core/Operations/CalculateVibration/FiniteElement/BeamWithDva/CalculateBeamWithDvaVibration.cs
@@ -2,8 +2,7 @@
 using IcVibracoes.Common.Classes;
 using IcVibracoes.Common.Profiles;
 using IcVibracoes.Core.ArrayOperations;
-using IcVibracoes.Core.AuxiliarOperations.BoundaryCondition;
-using IcVibracoes.Core.AuxiliarOperations.File;
+using IcVibracoes.Core.BoundaryCondition;
 using IcVibracoes.Core.Calculator.MainMatrixes.BeamWithDva;
 using IcVibracoes.Core.Calculator.NaturalFrequency;
 using IcVibracoes.Core.Calculator.Time;
@@ -44,7 +43,6 @@ namespace IcVibracoes.Core.Operations.CalculateVibration.FiniteElement.BeamWithD
         /// <param name="mappingResolver"></param>
         /// <param name="mainMatrix"></param>
         /// <param name="profileValidator"></param>
-        /// <param name="file"></param>
         /// <param name="time"></param>
         /// <param name="naturalFrequency"></param>
         public CalculateBeamWithDvaVibration(
@@ -54,10 +52,9 @@ namespace IcVibracoes.Core.Operations.CalculateVibration.FiniteElement.BeamWithD
             IMappingResolver mappingResolver,
             IBeamWithDvaMainMatrix<TProfile> mainMatrix,
             IProfileValidator<TProfile> profileValidator,
-            IFile file,
             ITime time,
             INaturalFrequency naturalFrequency)
-            : base(profileValidator, file, time, naturalFrequency)
+            : base(profileValidator, time, naturalFrequency)
         {
             this._boundaryCondition = boundaryCondition;
             this._arrayOperation = arrayOperation;